requirements.txt
bson==0.3.3
docutils==0.12
Flask==0.10.1
Geohash==1.0
gevent==1.0.1
greenlet==0.4.5
gunicorn==19.2.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
numpy==1.9.1
pandas==0.15.2
python-dateutil==2.4.0
python-memcached==1.53
pytz==2014.10
six==1.9.0
Werkzeug==0.9.6
